MDL-82584 mod_workshop: Refactor DB queries
Refactor workshop participant queries to improve performance.

This commit refactors how the authors/reviewers selectors are done in
mod_workshop. Previously this was trying to do all the work within SQL
so querying user roles and capabilities, with massive UNIONS for each
group on the course, which on big courses lead to SQL thousands of
lines long which broke the page, or took 5+ minutes to load.

This has been refactored to try and query only what is required
and then do capability checks within code instead of SQL.