MDL-84405 qbank_usage: Add filtering by question last used time.
This adds a new filter to the question bank to search by the last
time a question was used in a quiz.

It also adds behat tests and a change to the behat_mod_quiz helper
to allow us to set a quiz attempt at a specific time, to help with
testing last used dates in the past.